refactor: group style components in a package
