+++ studio/src/app/components/editor/styles/app-expansion-panel/app-expansion-panel.scss
