+++ studio/src/app/components/editor/styles/app-image/app-image.scss
