+++ studio/src/app/components/editor/styles/app-image/app-image.tsx
@@ -2,10 +2,10 @@ import {Component, Element, Prop, State, h, EventEmitter, Event} from '@stencil/
 
 import {alertController} from '@ionic/core';
 
-import {EditAction} from '../../../utils/editor/edit-action';
-import {ImageAction} from '../../../utils/editor/image-action';
+import {EditAction} from '../../../../utils/editor/edit-action';
+import {ImageAction} from '../../../../utils/editor/image-action';
 
-import {ImageHistoryService} from '../../../services/editor/image-history/image-history.service';
+import {ImageHistoryService} from '../../../../services/editor/image-history/image-history.service';
 
 @Component({
   tag: 'app-image',
