+++ studio/src/app/components/editor/styles/deck/app-deck-transition/app-deck-transition.scss
