+++ studio/src/app/components/editor/styles/element/app-align/app-align.scss
